:sparkles: Get multiple/defaultValue types working as intended
When multiple is set to true, the defaultValue is expected to be an
array of the base type of the component. When it's false, it should
be just the base type.

Note that this is not enabled by default for all input component
types, because the singular type for file field is already array
and the multiple flag only controls run-time behaviour but no changes
in the datatype of (default) values. In fact, defaultValue does not
even make any sense for file uploads.