:boom: [open-formulieren/open-forms#2958] Change types for component translation storage
Rather than storing translations on the entire form
definition, we will now track them on the component
definition itself. This results in what was internal
API/implementation detail to become public API, and
the types for this representation are now updated.

We store translations in openForms.translations for
each component, which itself is a mapping of
language code to a mapping of translations. Such a
translation mapping has the translatable properties
(of the component) as key and the translation as
value.