:fire: [#2] Drop (unintended) support for cosign registration attribute
This snuck in accidentally because the component inherits from the
email component in the implementation, but it was not a conscious
decision.

This does *not* make it impossible to send the data to StUF-ZDS, ZGW
APIs or Objects API integration as the unmapped data is included by
default in some form anyway or there are existing mechanisms to
include it via custom data templates.