:recycle: [open-formulieren/open-forms#4542] Refactor how openForms.* extensions are defined for a component type
Instead of having to re-define the openForms key itself and re-compose
it with the translatable keys, we can instead use a type variable to
specify the extensions to add via the OFExtensions helper type.

This removes the amount of possible errors and repetition w/r to the
translation keys that need to be passed, and the re-use of the generic
types is greater.