:sparkles: [open-formulieren/open-forms#4542] Added the optional requireVerification email extension
The extension is added under the openForms namespace to prevent
possible future collissions, and to make it explicit this is an
additional feature added by Open Forms.

The property is optional so that existing email components do not
require a data migration - not specifying it is falsy, which is the
default behaviour.

The property is not added under the 'validate' namespace, because it's
something that can only be properly validated server-side - it does not
make sense to configure client-side validation error messages for this
validation step because of that reason. We can still emit validation
errors server-side in a way that they will probably be displayed with
the component as if it was client-side validation.