:fire: [open-formulieren/open-forms#2177] Removing circle from the available map interactions
circle interaction is a leaflet custom thing, that isn't supported by geoJson. (To make it work, we would have to define the shape as a marker interaction, with extra custom properties).
Due to this reason, the registration apis don't/probably won't support this interaction.

For now, it would be better to not allow circle interactions