:label: [open-formulieren/formio-renderer#59] Correctly type the conditional.eq property
Formio (correctly) types this as a string for how their own SDK works,
which is by treating everything as a string and casting non-string
values to a string before comparing it to component.conditional.eq.

However, our own builder and backend actually require the correct JSON
types, e.g. a number component comparison has a number type for eq while
a checkbox component reference has a boolean type. We widen the type
here to match the behaviour of the builder and make it possible to
implement strong typed behaviour in our own renderer.

We achieve this by removing and re-adding the property from the
reference type of the Formio SDK, since extending/inheriting is not
allowed (you can only narrow types that way).